requirements.txt
fastapi
uvicorn
python-dotenv
httpx
supabase
python-telegram-bot>=20.0
nest_asyncio
loguru
tenacity
PyPDF2
python-multipart
pydantic-settings
openai
asyncpg
aiohttp
yandex-neural-api
numpy<2.0.0
pgvector>=0.4.2
pydub
scipy
tone @ git+https://github.com/voicekit-team/T-one.git
miniaudio